Added access control, users can't delete another users posts or edit them. Also added favicon.

+++ src/api/app/Http/Controllers/PostsController.php
@@ -8,6 +8,17 @@ use DB;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show'] ]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -90,6 +101,14 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        /**
+         * Checking if user is correct
+         */
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unathorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -126,8 +145,15 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
-
+        
+        /**
+         * Checking if user is correct
+         */
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unathorized Page');
+        }
+        
+        $post->delete();        
         return redirect('/posts')->with('success', 'Post Removed');
     }
 }
